+++ plugins/hazelcast3/pom.xml
@@ -16,7 +16,12 @@
       <dependency>
          <groupId>com.hazelcast</groupId>
          <artifactId>hazelcast</artifactId>
-         <version>3.0-RC1</version>
+         <version>3.0-RC2</version>
+      </dependency>
+      <dependency>
+         <groupId>org.radargun</groupId>
+         <artifactId>plugin-hazelcast2</artifactId>
+         <version>${project.version}</version>
       </dependency>
    </dependencies>
