+++ plugins/hazelcast3/src/main/resources/cacheprovider.properties
@@ -1,2 +1,2 @@
 ## This file is NECESSARY!
-org.radargun.wrapper org.radargun.cachewrappers.HazelcastWrapper
+org.radargun.wrapper org.radargun.cachewrappers.Hazelcast3Wrapper
